Modify runner.rs, tags for stages not parsed
The tops of the snippet qasm files contain directives controlling
which stages to run and what to expect. Previously, these were
parsed with the remainder of the code. This means changing the directives
forces running that stage again, causing a bigger diff than needed.

This commit strips the header before parsing